requirements.txt
Flask==2.3.2
pandas==2.1.1
python-docx==0.8.11
